pyproject.toml
[tool.poetry]
name = "beytv-replit"
version = "1.0.0"
description = "BeyTV Replit Edition - Lightweight Media Dashboard"
authors = ["BeyTV Team"]

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.32.3"
feedparser = "^6.0.11"
python-dotenv = "^1.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
